moved logic from ReactiveViewController into new class ReactiveGestureReactor
